refactor(manager): move `applyGitSource` to `util`
