+++ lib/modules/manager/pep621/processors/uv.ts
@@ -3,18 +3,12 @@ import { quote } from 'shlex';
 import { TEMPORARY_ERROR } from '../../../../constants/error-messages';
 import { logger } from '../../../../logger';
 import type { HostRule } from '../../../../types';
-import { detectPlatform } from '../../../../util/common';
 import { exec } from '../../../../util/exec';
 import type { ExecOptions, ToolConstraint } from '../../../../util/exec/types';
 import { getSiblingFileName, readLocalFile } from '../../../../util/fs';
-import { parseGitUrl } from '../../../../util/git/url';
 import { find } from '../../../../util/host-rules';
 import { Result } from '../../../../util/result';
 import { parseUrl } from '../../../../util/url';
-import { GitRefsDatasource } from '../../../datasource/git-refs';
-import { GitTagsDatasource } from '../../../datasource/git-tags';
-import { GithubTagsDatasource } from '../../../datasource/github-tags';
-import { GitlabTagsDatasource } from '../../../datasource/gitlab-tags';
 import { PypiDatasource } from '../../../datasource/pypi';
 import type {
   PackageDependency,
@@ -22,6 +16,7 @@ import type {
   UpdateArtifactsResult,
   Upgrade,
 } from '../../types';
+import { applyGitSource } from '../../util';
 import { type PyProject, UvLockfileSchema } from '../schema';
 import { depTypes, parseDependencyList } from '../utils';
 import type { PyProjectProcessor } from './types';
@@ -186,43 +181,6 @@ export class UvProcessor implements PyProjectProcessor {
         },
       ];
     }
-  }
-}
-
-function applyGitSource(
-  dep: PackageDependency,
-  git: string,
-  rev: string | undefined,
-  tag: string | undefined,
-  branch: string | undefined,
-): void {
-  if (tag) {
-    const platform = detectPlatform(git);
-    if (platform === 'github' || platform === 'gitlab') {
-      dep.datasource =
-        platform === 'github'
-          ? GithubTagsDatasource.id
-          : GitlabTagsDatasource.id;
-      const { protocol, source, full_name } = parseGitUrl(git);
-      dep.registryUrls = [`${protocol}://${source}`];
-      dep.packageName = full_name;
-    } else {
-      dep.datasource = GitTagsDatasource.id;
-      dep.packageName = git;
-    }
-    dep.currentValue = tag;
-    dep.skipReason = undefined;
-  } else if (rev) {
-    dep.datasource = GitRefsDatasource.id;
-    dep.packageName = git;
-    dep.currentDigest = rev;
-    dep.replaceString = rev;
-    dep.skipReason = undefined;
-  } else {
-    dep.datasource = GitRefsDatasource.id;
-    dep.packageName = git;
-    dep.currentValue = branch;
-    dep.skipReason = branch ? 'git-dependency' : 'unspecified-version';
   }
 }
